--- dongtai-core/src/main/java/io/dongtai/iast/core/bytecode/enhance/plugin/limiter/checker/impl/CpuUsageChecker.java
@@ -1,45 +1,0 @@
-package io.dongtai.iast.core.bytecode.enhance.plugin.limiter.checker.impl;
-
-import io.dongtai.iast.common.entity.performance.PerformanceMetrics;
-import io.dongtai.iast.common.entity.performance.metrics.CpuInfoMetrics;
-import io.dongtai.iast.common.entity.performance.metrics.MemoryUsageMetrics;
-
-import java.util.Properties;
-
-
-/**
- * cpu使用率检查器
- *
- * @author chenyi
- * @date 2022/3/4
- */
-public class CpuUsageChecker extends BasePerformanceChecker {
-
-    @Override
-    public boolean isPerformanceRisk(PerformanceMetrics nowMetrics, Properties cfg) {
-        final PerformanceMetrics thresholdMetrics = getMatchRiskThreshold(nowMetrics.getMetricsKey(), cfg);
-        if (thresholdMetrics != null) {
-            final CpuInfoMetrics threshold = thresholdMetrics.getMetricsValue(CpuInfoMetrics.class);
-            final CpuInfoMetrics now = nowMetrics.getMetricsValue(CpuInfoMetrics.class);
-            // cpu使用率
-            if (threshold.getCpuUsagePercentage() != null) {
-                return now.getCpuUsagePercentage() >= threshold.getCpuUsagePercentage();
-            }
-        }
-        return false;
-    }
-
-    @Override
-    public boolean isPerformanceOverLimit(PerformanceMetrics nowMetrics, Properties cfg) {
-        final PerformanceMetrics thresholdMetrics = getMatchMaxThreshold(nowMetrics.getMetricsKey(), cfg);
-        if (thresholdMetrics != null) {
-            final CpuInfoMetrics threshold = thresholdMetrics.getMetricsValue(CpuInfoMetrics.class);
-            final CpuInfoMetrics now = nowMetrics.getMetricsValue(CpuInfoMetrics.class);
-            // cpu使用率
-            if (threshold.getCpuUsagePercentage() != null) {
-                return now.getCpuUsagePercentage() >= threshold.getCpuUsagePercentage();
-            }
-        }
-        return false;
-    }
-}
