--- dongtai-core/src/main/java/io/dongtai/iast/core/bytecode/enhance/plugin/limiter/impl/AbstractRateLimiter.java
@@ -1,17 +1,0 @@
-package io.dongtai.iast.core.bytecode.enhance.plugin.limiter.impl;
-
-/**
- * 限速器抽象类
- *
- * @author liyuan40
- * @date 2022/3/10 15:06
- */
-public abstract class AbstractRateLimiter {
-
-    /**
-     * 获取令牌
-     *
-     * @return boolean 是否获取成功
-     */
-    public abstract boolean acquire();
-}
